+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Replace `babel-eslint` with `@babel/eslint-parser`. Refs UIOR-962.
 * Add cancel POL action and display indication that POL is canceled. Refs UIOR-740.
 * Update Order Permission. Refs UIOR-963.
+* Add organizations "Type" to order csv export. Refs UIOR-893.
 
 ## [3.1.2](https://github.com/folio-org/ui-orders/tree/v3.1.2) (2022-03-25)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v3.1.1...v3.1.2)